@@ -1,17 +1,7 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import App from './App.vue'
-import router from './router'
-import store from './store'
-import vuetify from './plugins/vuetify'
+import router from './router/index'
 import VueTheMask from 'vue-the-mask'
+import './index.css'
 
-
-Vue.use(VueTheMask)
-Vue.config.productionTip = false
-
-new Vue({
-  router,
-  store,
-  vuetify,
-  render: h => h(App)
-}).$mount('#app')
+createApp(App).use(router).use(VueTheMask).mount('#app')
