@@ -1,56 +1,20 @@
 <template>
-  <v-app>
-    <Navigation />
-    <v-main>
-      <router-view/>
-    </v-main>
+  <div class="overflow-x-hidden">
+    <Header />
+    <router-view />
     <Footer />
-  </v-app>
+  </div>
 </template>
 
 <script>
-import Navigation from './components/Navigation'
-import Footer from './components/Footer'
+import Header from "./views/Header/Header.vue";
+import Footer from "./components/Footer/Footer.vue";
 
 export default {
-  name: 'App',
-
-  data: () => ({
-    links: [
-      {nome: 'Home', path: "/"},
-      {nome: 'Feijões', path: "/"},
-      {nome: 'Fale Conosco', path: "/"},
-    ]
-  }),
-
-  computed: {
-    showTemplateTabs() {
-      switch (this.$vuetify.breakpoint.name) {
-        case "xs":
-          return false;
-        case "sm":
-          return false
-
-        default: return true;
-      }
-    }
-  },
-
   components: {
-    Navigation,
-    Footer
-  }
+    Header,
+    Footer,
+  },
 };
 </script>
 
-<style >
-  .center-me {
-  
-    position: absolute  !important;
-    margin-left: 50%;
-    transform: translate(-50%, 0);
-    z-index: 999999 !important;
-  }
-
-  
-</style>
