@@ -1,27 +1,17 @@
-import Vue from 'vue'
-import VueRouter from 'vue-router'
-import Index from '@/views/Index.vue'
-import Whatsapp from '@/views/Whatsapp.vue'
-
-Vue.use(VueRouter)
+import { createRouter, createWebHistory } from "vue-router";
+import IndexHome from "../views/Home/Index.vue";
+import IndexAbout from "../views/About/Index.vue";
+import IndexService from "../views/Service/Index.vue";
 
 const routes = [
-  {
-    path: '/',
-    name: 'Home',
-    component: Index
-  },
-  {
-    path: '/whatsapp',
-    name: 'Whatsapp',
-    component: Whatsapp
-  }
-]
+  { path: "/", name: "Home", component: IndexHome },
+  { path: "/about", name: "About", component: IndexAbout },
+  { path: "/services", name: "Service", component: IndexService },
+];
 
-const router = new VueRouter({
-  // mode: 'history',
-  base: process.env.BASE_URL,
-  routes
-})
+const router = createRouter({
+  history: createWebHistory(),
+  routes,
+});
 
-export default router
+export default router;
